add scores

+++ infrastructure/src/main/java/br/experian/com/adapters/PersonJpaAdapter.java
@@ -2,6 +2,7 @@ package br.experian.com.adapters;
 
 import br.experian.com.data.PersonDTO;
 import br.experian.com.entity.Person;
+import br.experian.com.exception.NotFoundException;
 import br.experian.com.ports.spi.PersonPersistencePort;
 import br.experian.com.repository.PersonRepository;
 import org.modelmapper.ModelMapper;
@@ -35,35 +36,43 @@ public class PersonJpaAdapter implements PersonPersistencePort {
         Optional<Person> optional = personRepository.findById(uuid);
         if (optional.isPresent()) {
             Person person = optional.get();
+            person.setName(entity.getName());
             person.setPhone(entity.getPhone());
             person.setAge(entity.getAge());
             person.setCity(entity.getCity());
+            person.setState(entity.getState());
             person.setScore(entity.getScore());
             person.setRegion(entity.getRegion());
 
             Person simulationSaved = personRepository.save(person);
             return modelMapper.map(simulationSaved, PersonDTO.class);
         }
-        return null;
+        else
+            throw new NotFoundException("person not found");
     }
 
     @Override
-    public List<PersonDTO> findAll() {
+    public List<PersonDTO> findAll() throws Exception {
         List<Person> personList = personRepository.findAll();
         Type listType = new TypeToken<List<PersonDTO>>(){}.getType();
         return modelMapper.map(personList, listType);
     }
 
     @Override
-    public PersonDTO findById(UUID entityId) {
+    public PersonDTO findById(UUID entityId) throws Exception {
         Optional<Person> personOptional = personRepository.findById(entityId);
         if (personOptional.isPresent())
             return modelMapper.map(personOptional.get(), PersonDTO.class);
-        return null;
+        else
+            throw new NotFoundException("person not found");
     }
 
     @Override
-    public void deleteById(UUID entityId) {
-        personRepository.deleteById(entityId);
+    public void deleteById(UUID entityId) throws Exception {
+        Optional<Person> personOptional = personRepository.findById(entityId);
+        if (personOptional.isPresent())
+            personRepository.deleteById(entityId);
+        else
+            throw new NotFoundException("person not found");
     }
 }
